Removing the 'Companies' component from the dashboard and moving to a screen
