+++ frontend/src/screens/dashboard/index.js
@@ -1,30 +1,13 @@
-import React, { PureComponent } from 'react'
-import { connect } from 'react-redux'
-import { bindActionCreators } from 'redux'
-import { getAllCompanies } from 'redux-flow/companies/reducer'
+import React from 'react'
 
 import Loading from 'components/loading'
 import AppLayout from 'components/layout'
-import Companies from './companies'
 
-class Dashboard extends PureComponent {
-  componentDidMount () {
-    const { getAllCompanies } = this.props
-    getAllCompanies()
-  }
+const Dashboard = ({ isFetching = false }) => [
+  <Loading key='loading' fetching={isFetching} />,
+  <AppLayout key='dashboard'>
+    Dashboard
+  </AppLayout>
+]
 
-  render () {
-    const { isFetching, companies } = this.props
-
-    return [
-      <Loading key='loading' fetching={isFetching} />,
-      <AppLayout key='dashboard'>
-        <Companies companies={companies} />
-      </AppLayout>
-    ]
-  }
-}
-
-const mapStateToProps = ({ companies }) => ({ ...companies })
-const mapDispatchToProps = dispatch => bindActionCreators({ getAllCompanies }, dispatch)
-export default connect(mapStateToProps, mapDispatchToProps)(Dashboard)
+export default Dashboard
